requirements.txt
fastapi
uvicorn
openai
pinecone-client==3.2.2
python-dotenv
python-multipart
pillow